+++ CHANGELOG.md
@@ -14,8 +14,9 @@ The format of this changelog is based on
   - Added `overlay_index` keyword to `terminate!` to allow applying terminations to overlay styles
   - Changed default CPW mesh size to use `2 * min(trace, gap)` (higher element quality when trace and gap are very different)
   - Changed default global mesh grading from `0.9` to `0.75` (more robust meshing for complex geometries, relatively small cost)
+  - Added `margin` keyword to `terminate!` to allow terminating a specified distance before the end of the path
   - Fixed incorrect behaviors when extending certain `Paths`: overlay styles continue as overlays, while terminations continue as `NoRenderContinuous`
-  - Fixed interactions between compound, decorated, overlay, and termination styles
+  - Fixed incompatibility issues for combinations of compound, decorated, overlay, and termination styles
   - Fixed bug where zero-length path segments could cause SolidModel rendering to fail
   - Fixed `SolidModel` rendering issue where some exterior boundaries might not be tagged
   - Fixed breaking error with `apply_size_to_surfaces=true` supplied via `MeshingParameters`; it is still deprecated as of 1.8.0 and has no effect, but no longer throws an error